todo: fix car->wait = op[car->op_code].cycle_wait;

+++ VM/srcs/update_pos.c
@@ -6,13 +6,13 @@
 /*   By: aelphias <aelphias@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/07 18:43:04 by aelphias          #+#    #+#             */
-/*   Updated: 2021/01/07 22:06:18 by aelphias         ###   ########.fr       */
+/*   Updated: 2021/01/08 21:29:37 by aelphias         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "corewar.h"
 
-uint8_t	update_pos(uint32_t pos)
+unsigned int	update_pos(unsigned int pos)
 {
 	if (pos >= MEM_SIZE)
 	pos = pos % MEM_SIZE;
